feat: implement layer management with commands for creating and deleting text and image layers

+++ Q2.Web-Service.API/DesignLab/Domain/Model/Commands/CreateImageLayerCommand.cs
@@ -3,4 +3,4 @@ using Q2.Web_Service.API.DesignLab.Domain.Model.ValueObjects;
 
 namespace Q2.Web_Service.API.DesignLab.Domain.Model.Commands;
 
-public record CreateImageLayerCommand(ProjectId ProjectId, String ImageUrl, float Width, float Height);
+public record CreateImageLayerCommand(ProjectId ProjectId, String ImageUrl, int Width, int Height);
